rename FirstRecordWaitTimeUtil to RecordWaitTimeUtil
